Fixed data types

+++ sys/com0com.h
@@ -19,6 +19,9 @@
  *
  *
  * $Log$
+ * Revision 1.7  2005/05/19 08:23:40  vfrolov
+ * Fixed data types
+ *
  * Revision 1.6  2005/05/14 17:07:02  vfrolov
  * Implemented SERIAL_LSRMST_MST insertion
  *
@@ -94,7 +97,7 @@ typedef struct _C0C_BUFFER {
   PUCHAR                  pBusy;
   PUCHAR                  pFree;
   PUCHAR                  pEnd;
-  ULONG                   busy;
+  SIZE_T                  busy;
   BOOLEAN                 escape;
   C0C_RAW_DATA            insertData;
 } C0C_BUFFER, *PC0C_BUFFER;
